Created page for each problem's user solutions
