--- src/pages/problem-solutions.tsx
@@ -1,45 +1,0 @@
-import { navigate, PageProps } from 'gatsby';
-import * as React from 'react';
-import ProblemSolutions from '../components/ProblemSolutions';
-import SubmitProblemSolutionModal from '../components/SubmitProblemSolutionModal';
-import { ProblemInfo } from '../models/problem';
-
-export default function ProblemSolutionsPage(
-  props: PageProps<null, null, { problem: ProblemInfo | null }>
-) {
-  const [isSubmitModalOpen, setIsSubmitModalOpen] = React.useState(false);
-
-  // Note: In the future, maybe this page should be generated by gatsby
-  // or we can import all the problems as a big JSON file or something.
-  const problem: ProblemInfo = props.location.state?.problem;
-
-  React.useEffect(() => {
-    if (!problem) navigate('/', { replace: true });
-  }, []);
-
-  if (!problem) {
-    return null;
-  }
-
-  const handleGoBack = () => {
-    navigate(-1);
-  };
-  const handleShowSubmitSolutionModal = () => {
-    setIsSubmitModalOpen(true);
-  };
-
-  return (
-    <>
-      <ProblemSolutions
-        onClose={handleGoBack}
-        showSubmitSolutionModal={handleShowSubmitSolutionModal}
-        problem={problem}
-      />
-      <SubmitProblemSolutionModal
-        problem={problem}
-        isOpen={isSubmitModalOpen}
-        onClose={() => setIsSubmitModalOpen(false)}
-      />
-    </>
-  );
-}
